requirements.txt
streamlit
Pillow
potracer